refactor: remove trigger & pipeline
