+++ package.json
@@ -13,10 +13,6 @@
       "types": "./lib/injection.d.ts",
       "default": "./lib/injection.js"
     },
-    "./pipeline": {
-      "types": "./lib/pipeline.d.ts",
-      "default": "./lib/pipeline.js"
-    },
     "./utils/*": {
       "types": "./lib/utils/*.d.ts",
       "default": "./lib/utils/*.js"
@@ -66,7 +62,6 @@
   },
   "dependencies": {
     "@artus/injection": "^0.5.1",
-    "@artus/pipeline": "^0.2.2",
     "deepmerge": "^4.2.2",
     "minimatch": "^5.0.1"
   },
