--- src/pipeline.ts
@@ -1,1 +1,0 @@
-export * from '@artus/pipeline';
