--- test/fixtures/app_koa_with_ts/src/http_trigger.ts
@@ -1,25 +1,0 @@
-import { Context, Next } from '@artus/pipeline';
-import { Injectable, ScopeEnum, Trigger } from '../../../../src';
-
-@Injectable({ scope: ScopeEnum.SINGLETON })
-export default class HttpTrigger extends Trigger {
-  constructor() {
-    super();
-    // first of all
-    this.use(async (ctx: Context, next: Next) => {
-      await next();
-      await this.respond(ctx);
-    });
-  }
-
-  async respond(ctx: Context) {
-    const { koaCtx } = ctx.input.params;
-    const { data } = ctx.output;
-
-    koaCtx.status = data.status || 200;
-    koaCtx.body = data.content;
-    for (const [k, v] of Object.entries(data.headers)) {
-      koaCtx.set(k, v);
-    }
-  }
-}
