+++ test/fixtures/app_koa_with_ts/src/lifecycle.ts
@@ -1,7 +1,6 @@
 import { DefaultContext } from 'koa';
 import { Server } from 'http';
-import { Container, Inject } from '@artus/injection';
-import { Context, Input } from '@artus/pipeline';
+import { Container, ExecutionContainer, Inject } from '@artus/injection';
 
 import { ArtusApplication, ArtusInjectEnum } from '../../../../src';
 import { LifecycleHookUnit, LifecycleHook } from '../../../../src/decorator';
@@ -9,7 +8,6 @@ import { ApplicationLifecycle } from '../../../../src/types';
 
 import KoaApplication from './koa_app';
 import HelloController from './controllers/hello';
-import HttpTrigger from './http_trigger';
 
 export let server: Server;
 
@@ -19,30 +17,23 @@ export default class MyLifecycle implements ApplicationLifecycle {
   app: ArtusApplication;
   @Inject()
   container: Container;
-  @Inject()
-  trigger: HttpTrigger;
 
   get koaApp(): KoaApplication {
     return this.container.get(KoaApplication);
   }
 
-  @LifecycleHook()
-  async didLoad() {
-    this.trigger.use(async (ctx: Context) => {
-      const { koaCtx } = ctx.input.params;
-      ctx.container.set({ id: 'headers', value: koaCtx.headers });
-      ctx.output.data = await ctx.container.get(HelloController).index();
-    });
-  }
-
   @LifecycleHook('willReady')
   setKoaMiddleware() {
     this.koaApp.use(async (koaCtx: DefaultContext) => {
-      const input = new Input();
-      const { req, res } = koaCtx;
-      input.params = { koaCtx, req, res };
-      const ctx = await this.trigger.initContext(input);
-      await this.trigger.startPipeline(ctx);
+      const executionContainer = new ExecutionContainer(null, this.container);
+      executionContainer.set({ id: 'headers', value: koaCtx.headers });
+
+      const data = await executionContainer.get(HelloController).index();
+      koaCtx.status = data.status || 200;
+      koaCtx.body = data.content;
+      for (const [k, v] of Object.entries(data.headers)) {
+        koaCtx.set(k, v);
+      }
     });
   }
