--- test/fixtures/trigger/event/event_trigger.ts
@@ -1,19 +1,0 @@
-import { Context, Next } from '@artus/pipeline';
-import { Injectable, ScopeEnum, Trigger } from '../../../../src';
-
-@Injectable({ scope: ScopeEnum.SINGLETON })
-export default class EventTrigger extends Trigger {
-  constructor() {
-    super();
-    // first of all
-    this.use(async (ctx: Context, next: Next) => {
-      await next();
-      await this.respond(ctx);
-    });
-  }
-
-  async respond(ctx: Context) {
-    const { output: { data: { type, payload } } } = ctx;
-    payload.cb(type);
-  }
-}
