--- test/fixtures/trigger/http/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "esModuleInterop": true,
-    "experimentalDecorators": true
-  },
-  "rootDir": "./",
-  "exclude": []
-}
